feat: add validation to utils
